Moved smile logic to new script, added smile anim trigger

+++ Smile Detector Test Task/Assets/FaceDetectionData.cs
@@ -12,7 +12,7 @@ namespace SmileDetectorTestTask
     [System.Serializable]
     public class FaceAttributes
     {
-        public double smile;
+        public float smile;
     }
 
     [System.Serializable]
